novas alteraçõese

+++ backend/src/controllers/DeleteBookController.ts
@@ -5,6 +5,14 @@ export class DeleteBookController {
   async handle(req: Request, res: Response) {
     const { id } = req.params;
 
+    const userRole = (req as any).user?.role;
+
+    if (userRole !== "ADMIN") {
+      return res.status(403).json({
+        error: "Acesso negado, somente admins podem deletar livros!"
+      });
+    }
+
     try {
       const service = new DeleteBookService();
       const result = await service.execute(Number(id));
